Refactor MCP generate rules from DTO tool
