+++ crates/rulemorph_mcp/src/tools/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod analyze_input;
 pub(crate) mod generate_rules_from_base;
+pub(crate) mod generate_rules_from_dto;
 pub(crate) mod list_ops;
